+++ app/Console/Commands/GenerateWorkflow.php
@@ -3,6 +3,7 @@
 namespace App\Console\Commands;
 
 use App\Objects\GuesserFiles;
+use App\Objects\ReportExecution;
 use App\Objects\WorkflowGenerator;
 use Illuminate\Auth\GenericUser;
 use Illuminate\Console\Command;
@@ -20,6 +21,7 @@ class GenerateWorkflow extends Command
      */
     protected $signature = 'ghygen:generate
     {--projectdir= : the directory of the project with composer.json}
+    {--dry-run : show only the execution report (no workflow file is create or showed)}
     {--cache : enable caching packages in the workflow}
     {--envfile=' . GuesserFiles::ENV_DEFAULT_TEMPLATE_FILE . ' : the .env file to use in the workflow}
     {--prefer-stable : Prefer stable versions of dependencies}
@@ -46,11 +48,15 @@ class GenerateWorkflow extends Command
         parent::__construct();
     }
 
-    private function printline(string $string, string $string2 = ""): void
+    public function title(string $title): void
     {
-        if ($this->saveFile) {
-            $this->line($string . " <info>" . $string2 . "</info>");
-        }
+        $this->line(str_pad("", strlen($title) + 12, "*"), "info");
+        $this->line("***   " . $title . "   ***", "info");
+        $this->line(str_pad("", strlen($title) + 12, "*"), "info");
+        $this->newLine();
+        $this->line("For auto generating the GitHub Actions Workflow,");
+        $this->line("I'm going to analyze the project requirements...");
+        $this->newLine();
     }
 
     /**
@@ -60,10 +66,14 @@ class GenerateWorkflow extends Command
      */
     public function handle()
     {
+        $reportExecution = new ReportExecution();
         $this->saveFile = false;
+        $dryRun = $this->option("dry-run");
         $projectdir = $this->option("projectdir");
         if (is_null($projectdir)) {
             $projectdir = "";
+        } else {
+            $projectdir = rtrim($projectdir, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR;
         }
         $yamlFile = $this->option("save");
         $this->saveFile = ! is_null($yamlFile);
@@ -92,11 +102,11 @@ class GenerateWorkflow extends Command
         $generator = new WorkflowGenerator();
         $generator->loadDefaults();
 
-        if ($guesserFiles->composerExists()) {
-            $this->printline("Composer file loaded");
+        if ($guesserFiles->composerExists()) { /** @phpstan-ignore-line */
+            $reportExecution->addValueInfo("Composer file", "Loaded");
             $composer = json_decode(file_get_contents($guesserFiles->getComposerPath()), true);
             $generator->name = Arr::get($composer, 'name', "");
-            $this->printline("Project name", $generator->name);
+            $reportExecution->addValueInfo("Project name", $generator->name);
             $yamlFile = GuesserFiles::generateYamlFilename(
                 GuesserFiles::getGithubWorkflowDirectory($projectdir),
                 $generator->name
@@ -105,7 +115,9 @@ class GenerateWorkflow extends Command
             $phpversion = Arr::get($composer, 'require.php', "8.0");
 
             $stepPhp = $generator->detectPhpVersion($phpversion);
-            $this->printline("PHP versions", implode(",", $stepPhp));
+            $reportExecution->addValueInfo("PHP versions", $stepPhp);
+
+            $optionPreferSelected = $this->option("prefer-stable") || $this->option("prefer-lowest");
             if ($this->option("prefer-stable") && $this->option("prefer-lowest")) {
                 $generator->dependencyStability = [ 'prefer-stable', 'prefer-lowest' ];
             } elseif ($this->option("prefer-lowest")) {
@@ -115,7 +127,11 @@ class GenerateWorkflow extends Command
             } else {
                 $generator->dependencyStability = [ 'prefer-none' ];
             }
-            $this->printline("Dependency stability", implode(",", $generator->dependencyStability));
+            if ($optionPreferSelected) {
+                $reportExecution->addValueInfo("Dependency stability", $generator->dependencyStability);
+            } else {
+                $reportExecution->addValueComment("Dependency stability", "Not selected");
+            }
 
             // detect packages
             $devPackages = Arr::get($composer, 'require-dev');
@@ -125,14 +141,18 @@ class GenerateWorkflow extends Command
                 $laravelVersions = GuesserFiles::detectLaravelVersionFromTestbench($testbenchVersions);
                 $generator->matrixLaravel = true;
                 $generator->matrixLaravelVersions = $laravelVersions;
-                $this->printline("Laravel versions", implode(",", $laravelVersions));
+                $reportExecution->addValueInfo("Laravel versions", $laravelVersions);
+            } else {
+                $reportExecution->addValueComment("Laravel", "No testbench detected");
             }
             // squizlabs/php_codesniffer
             $phpCodesniffer = Arr::get($devPackages, "squizlabs/php_codesniffer", "");
             if ($phpCodesniffer !== "") {
                 $generator->stepExecuteCodeSniffer = true;
                 $generator->stepInstallCodeSniffer = false;
-                $this->printline("Code sniffer", "Install");
+                $reportExecution->addValueInfo("Code sniffer", "Detected");
+            } else {
+                $reportExecution->addValueComment("Code sniffer", "Not detected");
             }
             // nunomaduro/larastan
             $larastan = Arr::get($devPackages, "nunomaduro/larastan", "");
@@ -141,7 +161,7 @@ class GenerateWorkflow extends Command
                 $generator->stepInstallStaticAnalysis = false;
                 $generator->stepToolStaticAnalysis = "larastan";
                 $generator->stepPhpstanUseNeon = $guesserFiles->phpstanNeonExists();
-                $this->printline("Static code analysis", "Larastan and PHPStan");
+                $reportExecution->addValueInfo("Static code analysis", "Larastan and PHPStan");
             } else {
                 $phpstan = Arr::get($devPackages, "phpstan/phpstan", "");
                 if ($phpstan !== "") {
@@ -149,7 +169,9 @@ class GenerateWorkflow extends Command
                     $generator->stepInstallStaticAnalysis = false;
                     $generator->stepToolStaticAnalysis = "phpstan";
                     $generator->stepPhpstanUseNeon = $guesserFiles->phpstanNeonExists();
-                    $this->printline("Static code analysis", "PHPStan");
+                    $reportExecution->addValueInfo("Static code analysis", "PHPStan");
+                } else {
+                    $reportExecution->addValueComment("Static code analysis", "Not detected");
                 }
             }
             $generator->stepDusk = false;
@@ -158,17 +180,28 @@ class GenerateWorkflow extends Command
             $phpunit = Arr::get($devPackages, "phpunit/phpunit", "");
             if ($phpunit !== "") {
                 $generator->stepExecutePhpunit = true;
-                $this->printline("Automated test", "PHPUnit");
+                $reportExecution->addValueInfo("Automated test", "PHPUnit");
             }
             // phpunit/phpunit
             $generator->stepExecutePestphp = false;
             $pestphp = Arr::get($devPackages, "pestphp/pest", "");
             if ($pestphp !== "") {
                 $generator->stepExecutePestphp = true;
-                $this->printline("Automated test", "Pest");
+                $reportExecution->addValueInfo("Automated test", "PestPHP");
             }
+            if ($pestphp === "" && $phpunit === "") {
+                $reportExecution->addValueComment("Automated test", "Not detected");
+            }
+        } else {
+            $reportExecution->addValueComment("Composer file", "not found");
         }
+
         $generator->detectCache($cache);
+        if ($cache) {
+            $reportExecution->addValueInfo("Caching packages", "Yes cache");
+        } else {
+            $reportExecution->addValueComment("Caching packages", "No cache");
+        }
 
         $generator->databaseType = WorkflowGenerator::DB_TYPE_NONE;
         $generator->stepRunMigrations = false;
@@ -182,22 +215,26 @@ class GenerateWorkflow extends Command
             $generator->stepRunMigrations = false;
             if ($databaseType === "mysql") {
                 $generator->databaseType = WorkflowGenerator::DB_TYPE_MYSQL;
-                $this->printline("Detected Mysql", "will setup Mysql service");
+                $reportExecution->addValueInfo("Database", "Mysql service");
             }
             if ($databaseType === "sqlite") {
                 $generator->databaseType = WorkflowGenerator::DB_TYPE_SQLITE;
-                $this->printline("Detected Sqlite", "done");
+                $reportExecution->addValueInfo("Database", "SQLite");
             }
             if ($databaseType === "postgresql") {
                 $generator->databaseType = WorkflowGenerator::DB_TYPE_POSTGRESQL;
-                $this->printline("Detected PostgreSQL", "will setup pgsql service");
+                $reportExecution->addValueInfo("Database", "PostgreSQL service");
             }
             if ($generator->databaseType !== WorkflowGenerator::DB_TYPE_NONE) {
                 $migrationFiles = scandir($guesserFiles->getMigrationsPath());
                 if (count($migrationFiles) > 4) {
                     $generator->stepRunMigrations = true;
-                    $this->printline("I will execute also migrations", "done");
+                    $reportExecution->addValueInfo("Migrations", "Detected");
+                } else {
+                    $reportExecution->addValueComment("Migrations", "Not detected");
                 }
+            } else {
+                $reportExecution->addValueComment("Database", "No database");
             }
         }
         if ($guesserFiles->packageExists()) {
@@ -208,20 +245,27 @@ class GenerateWorkflow extends Command
             if ($versionFromNvmrc !== "") {
                 $generator->stepNodejsVersion = $versionFromNvmrc;
             }
-            $this->printline("NodeJS detected", "version " .  $generator->stepNodejsVersion);
+            $reportExecution->addValueInfo("NodeJS/Npm", "Detected, version " .  $generator->stepNodejsVersion);
+        } else {
+            $reportExecution->addValueComment("NodeJS/Npm", "Not Detected");
         }
         $appKey = "";
         $generator->stepGenerateKey = false;
 
         if ($guesserFiles->envDefaultTemplateExists()) {
             $generator->stepCopyEnvTemplateFile = true;
+            $reportExecution->addValueInfo(".env file", "Detected");
             $generator->stepEnvTemplateFile = $optionEnvWorkflowFile;
             // Generate Key
             $envArray = $generator->readDotEnv($guesserFiles->getEnvDefaultTemplatePath());
             $appKey = Arr::get($envArray, "APP_KEY", "");
 
             $generator->stepGenerateKey = $appKey === "";
+            if ($generator->stepGenerateKey) {
+                $reportExecution->addValueInfo("Laravel App Key", "Will be generated");
+            }
         } else {
+            $reportExecution->addValueComment(".env file", "Not detected");
             $generator->stepCopyEnvTemplateFile = false;
         }
         $generator->stepFixStoragePermissions = false;
@@ -230,27 +274,42 @@ class GenerateWorkflow extends Command
             // fix storage permissions
             $generator->stepFixStoragePermissions = true;
         }
+        if ($generator->stepFixStoragePermissions) {
+            $reportExecution->addValueInfo("Laravel Fix storage permission", "Chmod will be executed");
+        }
 
 
         $data = $generator->setData();
 
         $result = $generator->generate($data);
-        if ($this->saveFile) {
-            try {
-                $size = file_put_contents($yamlFile, $result);
-                $this->info("File " . $yamlFile . " saved (" . $size . " bytes)");
-            } catch (\Exception $e) {
-                if (! GuesserFiles::existsGithubWorkflowDirectory($projectdir)) {
-                    $this->error(
-                        "Workflow directory doesn't exist : " .
-                        GuesserFiles::getGithubWorkflowDirectory($projectdir)
-                    );
-                } else {
-                    $this->error($e->getMessage());
-                }
-            }
+        if ($dryRun) {
+            $this->title("Ghygen");
+            $this->table(["Report", "Status"], $reportExecution->toArrayLabelValue());
         } else {
-            $this->line($result);
+            if ($this->saveFile) {
+                try {
+                    $size = file_put_contents($yamlFile, $result);
+                    $this->title("Ghygen");
+                    $this->table(["Report", "status"], $reportExecution->toArrayLabelValue());
+                    $this->info("File " . $yamlFile);
+                    $this->info("File created (" . $size . " bytes)");
+                } catch (\Exception $e) {
+                    if (! GuesserFiles::existsGithubWorkflowDirectory($projectdir)) {
+                        $this->error(
+                            "Workflow directory doesn't exist : " .
+                            GuesserFiles::getGithubWorkflowDirectory($projectdir)
+                        );
+                        $this->info(
+                            "Hint: create Workflow directory: " .
+                            GuesserFiles::getGithubWorkflowDirectory($projectdir)
+                        );
+                    } else {
+                        $this->error($e->getMessage());
+                    }
+                }
+            } else {
+                $this->line($result);
+            }
         }