+++ app/Providers/RouteServiceProvider.php
@@ -24,7 +24,6 @@ class RouteServiceProvider extends ServiceProvider
      *
      * When present, controller route declarations will automatically be prefixed with this namespace.
      *
-     * @var string|null
      */
     // protected $namespace = 'App\\Http\\Controllers';