+++ resources/views/components/form/input-conditional-checkbox.blade.php
@@ -7,7 +7,7 @@
     'model' => '',
 ])
 
-<div class="pl-3 pb-2 mt-2 space-y-4"  x-data="{ show{{ $id }}: @entangle($attributes->wire('model')) }">
+<div class="card bordered shadow pl-3 pb-2 m-2 space-y-4  "  x-data="{ show{{ $id }}: @entangle($attributes->wire('model')) }">
     <!-- div class="pl-3 pb-2 mt-2 space-y-4"-->
     <div>
         <x-form.input-checkbox
@@ -20,7 +20,7 @@
         </x-form.input-checkbox>
     </div>
 
-    <div class="pl-5 pt-2 pb-2 pr-5 bg-blue-50 rounded-2xl" x-show="show{{ $id }}">
+    <div class="" x-show="show{{ $id }}">
         {{ $slot }}
 
     </div>