+++ tailwind.config.js
@@ -12,7 +12,9 @@ module.exports = {
     extend: {},
   },
   plugins: [
-      require('@tailwindcss/forms'),
+
       require('@tailwindcss/typography'),
+      require('daisyui'),
+
   ],
 }